Creating separate .py file for NEA API to organize usage across multiple files. Removed constants.py as the majority of constants were being used for NEA.
