--- exotic/constants.py
@@ -1,49 +1,0 @@
-# ########################################################################### #
-#    Copyright (c) 2019-2020, California Institute of Technology.
-#    All rights reserved.  Based on Government Sponsored Research under
-#    contracts NNN12AA01C, NAS7-1407 and/or NAS7-03001.
-#
-#    Redistribution and use in source and binary forms, with or without
-#    modification, are permitted provided that the following conditions
-#    are met:
-#      1. Redistributions of source code must retain the above copyright
-#         notice, this list of conditions and the following disclaimer.
-#      2. Redistributions in binary form must reproduce the above copyright
-#         notice, this list of conditions and the following disclaimer in
-#         the documentation and/or other materials provided with the
-#         distribution.
-#      3. Neither the name of the California Institute of
-#         Technology (Caltech), its operating division the Jet Propulsion
-#         Laboratory (JPL), the National Aeronautics and Space
-#         Administration (NASA), nor the names of its contributors may be
-#         used to endorse or promote products derived from this software
-#         without specific prior written permission.
-#
-#    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-#    "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-#    LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-#    A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE CALIFORNIA
-#    INSTITUTE OF TECHNOLOGY BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-#    SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
-#    TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
-#    PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
-#    LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
-#    NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-#    SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-#
-# ########################################################################### #
-#    EXOplanet Transit Interpretation Code (EXOTIC)
-#    # NOTE: See companion file version.py for version info.
-# ########################################################################### #
-from numpy import pi
-import astropy.constants as const
-import astropy.units as u
-
-AU = const.au                                                       # m
-PI = pi
-R_SUN = const.R_sun                                                 # m
-R_JUP = const.R_jup                                                 # m
-
-# CALCULATED VALUES
-G = const.G.to(AU**3 / (const.M_sun * u.day**2))                    # AU^3 /(msun * day^2)
-SA = lambda m, p: (G * m * p ** 2. / (4. * PI ** 2.)) ** (1. / 3.)  # Keplerian semi-major axis (au)
